Controller tests

See merge request scrum_team_3/idatt2106_2024_03_backend!148

+++ src/main/java/idatt2106/systemutvikling/sparesti/service/ChallengeService.java
@@ -2,13 +2,18 @@ package idatt2106.systemutvikling.sparesti.service;
 
 import idatt2106.systemutvikling.sparesti.dao.ChallengeDAO;
 import idatt2106.systemutvikling.sparesti.dao.ChallengeLogDAO;
+import idatt2106.systemutvikling.sparesti.dao.MilestoneDAO;
 import idatt2106.systemutvikling.sparesti.dto.ChallengeDTO;
+import idatt2106.systemutvikling.sparesti.exceptions.BankConnectionErrorException;
+import idatt2106.systemutvikling.sparesti.exceptions.NotFoundInDatabaseException;
 import idatt2106.systemutvikling.sparesti.mapper.ChallengeMapper;
 import idatt2106.systemutvikling.sparesti.repository.ChallengeLogRepository;
 import idatt2106.systemutvikling.sparesti.repository.ChallengeRepository;
 
 import java.util.logging.Logger;
 
+import idatt2106.systemutvikling.sparesti.repository.MilestoneRepository;
+import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.*;
 import org.springframework.stereotype.Service;
@@ -25,8 +30,11 @@ import java.util.List;
 public class ChallengeService {
 
   private final Logger logger = Logger.getLogger(ChallengeService.class.getName());
-  private ChallengeRepository challengeRepository;
-  private ChallengeLogRepository challengeLogRepository;
+  private final TransactionService transactionService;
+  private final ChallengeRepository challengeRepository;
+  private final ChallengeLogRepository challengeLogRepository;
+  private final MilestoneService milestoneService;
+  private final MilestoneRepository dbMilestone;
 
   /**
    * Method to get a challenge by its id from the database and return it as a ChallengeDTO object to
@@ -208,20 +216,74 @@ public class ChallengeService {
     return challengeRepository.save(challengeDAO);
   }
 
+  @Transactional
+  public void completeChallengeForCurrentUser(long challengeId, long milestoneId) {
+    // Fetch username from security context
+    final String username = CurrentUserService.getCurrentUsername();
+
+    // Fetch data
+    ChallengeDAO challenge = challengeRepository.findByChallengeIdAndUserDAO_Username(challengeId, username);
+    MilestoneDAO milestone = dbMilestone.findMilestoneDAOByMilestoneIdAndUserDAO_Username(milestoneId, username);
+
+    // Verify existence of the requested challenge
+    if (challenge == null)
+      throw new NotFoundInDatabaseException("No challenge found with the requested ID for the user");
+
+    // Verify existence of the requested milestone
+    if (milestone == null)
+      throw new NotFoundInDatabaseException("No milestone found with the requested ID for the user");
+
+
+
+    // Define transfer amount as the difference between the goal and the current sum
+    long transferAmount = challenge.getGoalSum() - challenge.getCurrentSum();
+
+    // Perform savings transaction
+    boolean success = transactionService.createSavingsTransferForCurrentUser(transferAmount);
+
+    // Verify transaction success
+    if (success)
+      throw new BankConnectionErrorException("Failed to transfer funds to savings");
+
+
+    // Transfer achieved currency to milestone
+    MilestoneDAO savedEntry = milestoneService.increaseMilestonesCurrentSum(milestoneId, challenge.getGoalSum());
+
+
+    // Archive challenge
+    archiveActiveChallenge(challenge.getChallengeId());
+  }
+
   /**
-   * Method to complete a challenge. The method returns a ChallengeLogDAO object. The challenge is
-   * completed with the current date and time. The challenge is completed with the same information
-   * as the challenge. The challenge is deleted from the database and saved as a challenge log.
+   * Performs the changes - local to the challenge repository - needed for a challenge to be considered as 'completed'.
+   * This function creates a log entry from the selected challenge (specified by the id-parameter)
+   * and stores it in the database. The challenge is deleted from the table of active challenges,
+   * but the log entry persists.
    *
    * @param challengeId the id of the challenge to complete
-   * @return the completed challenge
+   * @return The log entry that is persisted by this function.
    */
-  public ChallengeLogDAO completeChallenge(Long challengeId) {
+  public ChallengeLogDAO archiveActiveChallenge(Long challengeId) {
+    // Fetch active challenge from database
     ChallengeDAO challengeDAO = challengeRepository.findChallengeDAOByChallengeId(challengeId);
+
+    // Verify existence of the active challenge
+    if (challengeDAO == null)
+      return null;
+
+    // Create log entry from the active challenge
     ChallengeLogDAO challengeLogDAO = createChallengeLog(challengeDAO);
+
+    // Set the archived currency equal to the goal of the active challenge
     challengeLogDAO.setChallengeAchievedSum(challengeLogDAO.getGoalSum());
+
+    // Save challenge log entry to database
+    challengeLogDAO = challengeLogRepository.save(challengeLogDAO);
+
+    // Remove active challenge from database
     challengeRepository.delete(challengeDAO);
-    challengeLogRepository.save(challengeLogDAO);
+
+    // Return the archived entry
     return challengeLogDAO;
   }
